矩形下载自动递归拆分

+++ BaiduMapFinder/LocationControl.xaml.cs
@@ -31,8 +31,8 @@ namespace BaiduMapFinder
 
         public class LocationModel
         {
-            public float? Lat { get; set; }
-            public float? Lng { get; set; }
+            public double? Lat { get; set; }
+            public double? Lng { get; set; }
         }
     }
 }
